# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Him For You discovery and membership API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registration/admin/registrations/{id}/review:
    {
      "post":
        {
          "operationId": "reviewRegistration",
          "tags": ["registration"],
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "string" },
              },
            ],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationReview" },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationRecord",
                            },
                        },
                    },
                },
            },
        },
    }
  /registration/profile:
    {
      "put":
        {
          "operationId": "updateMyRegistration",
          "tags": ["registration"],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationRecord" },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Updated private profile",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationRecord",
                            },
                        },
                    },
                },
            },
        },
    }
  /registration/admin/login:
    {
      "post":
        {
          "operationId": "loginRegistrationAdmin",
          "tags": ["registration"],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "$ref": "#/components/schemas/RegistrationAdminLogin",
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Saved",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationSettings",
                            },
                        },
                    },
                },
            },
        },
    }
  /registration/admin/settings:
    {
      "put":
        {
          "operationId": "updateRegistrationSettings",
          "tags": ["registration"],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationSettings" },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Saved",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationSettings",
                            },
                        },
                    },
                },
            },
        },
    }
...
  /registrations:
    {
      "post":
        {
          "operationId": "registerProfile",
          "tags": ["registration"],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationRecord",
                            },
                        },
                    },
                },
            },
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationInput" },
                    },
                },
            },
        },
    }
...
  /registration/login:
    {
      "post":
        {
          "operationId": "loginRegistration",
          "tags": ["registration"],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationRecord",
                            },
                        },
                    },
                },
            },
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationLogin" },
                    },
                },
            },
        },
    }
...
components:
...
  schemas:
...
    RegistrationPhoto:
      {
        "type": "object",
        "properties":
          {
            "dataUrl":
              { "type": "string", "minLength": 1, "maxLength": 1500000 },
            "category":
              {
                "type": "string",
                "enum": ["profile", "additional", "gallery"],
              },
          },
        "required": ["dataUrl", "category"],
      }
...
    RegistrationRecord:
      {
        "type": "object",
        "properties":
          {
            "displayName":
              { "type": "string", "minLength": 1, "maxLength": 80 },
            "dateOfBirth":
              { "type": "string", "minLength": 1, "maxLength": 10 },
            "email": { "type": "string", "minLength": 1, "maxLength": 254 },
            "mobile": { "type": "string", "minLength": 1, "maxLength": 25 },
            "headline": { "type": "string", "minLength": 0, "maxLength": 120 },
            "about": { "type": "string", "minLength": 0, "maxLength": 2000 },
            "country": { "type": "string", "minLength": 0, "maxLength": 80 },
            "state": { "type": "string", "minLength": 0, "maxLength": 80 },
            "city": { "type": "string", "minLength": 0, "maxLength": 80 },
            "area": { "type": "string", "minLength": 0, "maxLength": 100 },
            "age": { "type": "number", "minimum": 18, "maximum": 100 },
            "pinCode": { "type": "string", "minLength": 0, "maxLength": 12 },
            "interests":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "languages":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "preferences":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "availability":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "minAge": { "type": "number", "minimum": 18, "maximum": 100 },
            "maxAge": { "type": "number", "minimum": 18, "maximum": 100 },
            "status":
              { "type": "string", "enum": ["available", "unavailable"] },
            "listing":
              {
                "type": "string",
                "enum": ["free", "weekly", "monthly", "quarterly"],
              },
            "partnerOptIn": { "type": "boolean" },
            "accurate": { "type": "boolean" },
            "terms": { "type": "boolean" },
            "adult": { "type": "boolean" },
            "photos":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 8,
                "items": { "$ref": "#/components/schemas/RegistrationPhoto" },
              },
            "mainPhoto": { "type": "number", "minimum": 0, "maximum": 7 },
            "id": { "type": "string", "minLength": 1, "maxLength": 120 },
            "submittedAt":
              { "type": "string", "minLength": 1, "maxLength": 120 },
            "reviewStatus":
              { "type": "string", "minLength": 1, "maxLength": 120 },
            "listingPrice": { "type": "number", "minimum": 0 },
            "isComplete": { "type": "boolean" },
            "isPublished": { "type": "boolean" },
            "missingSteps": { "type": "array", "items": { "type": "string" } },
            "notificationStatus": { "type": "string" },
            "publicSlug": { "type": "string" },
          },
        "required":
          [
            "displayName",
            "dateOfBirth",
            "email",
            "mobile",
            "headline",
            "about",
            "country",
            "state",
            "city",
            "area",
            "age",
            "pinCode",
            "interests",
            "languages",
            "preferences",
            "availability",
            "minAge",
            "maxAge",
            "status",
            "listing",
            "partnerOptIn",
            "accurate",
            "terms",
            "adult",
            "photos",
            "mainPhoto",
            "id",
            "submittedAt",
            "reviewStatus",
            "listingPrice",
          ],
      }
    RegistrationSettings:
      {
        "type": "object",
        "properties":
          {
            "plans":
              {
                "type": "array",
                "items": { "$ref": "#/components/schemas/RegistrationPlan" },
              },
            "termsUrl": { "type": "string" },
            "privacyUrl": { "type": "string" },
          },
        "required": ["plans", "termsUrl", "privacyUrl"],
      }
    RegistrationPlan:
      {
        "type": "object",
        "properties":
          {
            "id":
              { "type": "string", "enum": ["weekly", "monthly", "quarterly"] },
            "price": { "type": "number", "minimum": 0 },
            "enabled": { "type": "boolean" },
          },
        "required": ["id", "price", "enabled"],
      }
    RegistrationLogin:
      {
        "type": "object",
        "properties":
          {
            "email": { "type": "string", "minLength": 1, "maxLength": 254 },
            "password": { "type": "string", "minLength": 1, "maxLength": 128 },
          },
        "required": ["email", "password"],
      }
    RegistrationAdminLogin:
      {
        "type": "object",
        "required": ["password"],
        "properties":
          {
            "password": { "type": "string", "minLength": 1, "maxLength": 256 },
          },
      }
    RegistrationReview:
      {
        "type": "object",
        "required": ["action"],
        "properties":
          { "action": { "type": "string", "enum": ["approve", "reject"] } },
      }