# lib/api-zod/src/generated/types/registrationLogin.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Him For You discovery and membership API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationLogin {
  /**
     * @minLength 1
     * @maxLength 254
     */
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}