# lib/api-zod/src/generated/types/registrationReview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Him For You discovery and membership API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationReviewAction } from './registrationReviewAction';

export interface RegistrationReview {
  action: RegistrationReviewAction;
}